requirements.txt
Flask
psycopg2-binary
gunicorn
flask-wtf
python-dotenv